Use absolute path to identify sources in compiler cache
This is required to avoid issues if a user is using multiple versions 
of the application on the same system. If a user installed two 
version of the application and ran them one ofter the other, the 
second would use the cached kernels of the first, as the filenames of 
the sources matched and the last modification of the cache file was 
after the creation of the source file.