Fix parsing of input parameters
The bug showed in the su3vec tests if the build was performed in 
release mode. There were basically two problems: * Option groups were 
returned by value from the get_option methods of the classes managing 
them. Thus a temporary was handed to options_description::add(), 
which expects a reference. It was thus holding a reference to a 
temporary variable, which was asking for trouble. * All the parameter 
groups were derived from the same abstract class, but again parents 
of the Inputparameters class. Such diamond shaped class diagrams 
require especial care (virtual parent relationship), which was not 
taken.