Changed constructors of Spinorfield and Spinorfield_eo
classes such that they accept suitable parametersinterfaces. 
Accordingly a lot of function signatures had to be adapted wherever 
interfaces are needed. Additionally in some functions new template 
functions/classes had to be introduced as workaround wherever 
template objects were instantiated for which it is not clear at 
compile time whether they need an interface or not.