Implemented new parameter for pseudofermions
I took the opportunity to refactor some code in the RHMC parameters 
class. In particular I transformed the int variables to unsigned 
where it made sense. This propagated a bit around to avoid 
signed-unsiged comparison. I also used camel case notation for the 
members of the class (still not for methods).