Work on hooks to check header of files
Files with extensions .bash .c .C .cl .cmake .cpp .h .hpp .m .nb.py .sh
are checked looking for the correct license header, which was added as
standalone in a file in the LICENSE folder. If it is not (completely)
found, the user is asked about what to do. This prompt allows to add
third party code which should not contain CL2QCD license header.