Change format of hmc and rhmc output files
The format (e.g. number of digits) of each field in the output files
has been fixed in a way that each line looks the same and no horizontal
shift occurs. The time per trajectory has been introduced in the output
files of the generation executables and unnecessary columns have been
removed (for example only DH and not exp(DH) or the acceptance
probability is printed).