Fix microbenchmark code
At the time of the big hardware refactoring (remove dependency on meta
package and refactor all tests with mockup objects) it was not noticed
that the code of the microbenchmark had been broken, probably because
it was not switched on in compilation. I fixed now the compilation
adding a method to the system class to get the kernel
parameters. Probably this is not ideal, since this method is used only
in the microbenchmarks. A comment was left about this.