Improve thermalisation in generation executable
During thermalisation, if implemented, it makes sense to measure the
gauge observables, which are usually cheap to measure and which can
make the user check about actual thermalisation afterwards.
An actual throw is done if the thermalisation is not implemented
(before just a warning was given and the algorithm was going on).
Some variables were also renamed with more meaningful names
and the writing to file of gauge observables was done in as in
the (R)HMC algortihms.