Refactor device code to semplify its structure
The more a kernel code is complicated, the higher is the probability
that the OpenCL compiler miscompiles it with consequent errors in
tests. The even-odd preconditioned Wilson code was implemented in a
simpler way with respect to non even-odd preconditioned code. The
latter was brought in an analogous form to the former. Please, note
that the AOS/SOA automatic switch is still not implemented in non
even-odd preconditione Wilson code.