Adjust boost test for more recent boost versions
At the moment we require version 1.59 of boost but in version 1.60
there was a breaking change, forcing the user to separate boost
arguments from user arguments with a '--'. This implies that in
versions prior to 1.60 the '--' if given is considered as user
argument, while from version 1.60 on it is not the case. Waiting to
require at least version 1.60 of boost in the future (the full unit
tests were already prepared for that), an if case on the boost version
was introduced to make the test pass with all boost versions.