Remove unnecessary user declared copy constructor
For a simple class like ProfilingData having only a couple of uint64 as
members, it should be enough the default copy constructor provided by
the compiler. A user defined one was provided but then also a copy
assignment operator was used in tests, which was causing a warning
since not provided.