Require aligment of double/float in kernels only
Variables attributes in template argument are ignored and with recent
compiler versions a huge amount of warnings were given because the
aligment in std::vector<hmc_float> or in Plain<hmc_float> could not be
ensured by the compiler. That aligment on host code should not play a
role and no amplification comment was left there by Matthias Bach. It
is reasonable to assume that such an alignment might be relevant in
respect to performance when compiling kernel code and therefore it has
been specified only in that case. Benchmarks need to be done on new
generation cards in any case and this will be reverted in case of
unexpected results.