Make behaviour about valid found devices uniform
The code is now forbidding using OpenCL devices of different types at
the same time and, in particular, if a GPU is selected, all CPUs are
excluded. The user can request either to use all available
devices (this is the default behaviour) or to use the first n of the
selected type or some devices by index, but in all case mixing
different types is forbidden and GPUs have priority on CPUs.