Make tests use any device if nothing is specified
In particular, if not differently specified via command line options,
the System object for the tests will be initialized with both CPUs and
GPUs as valid device and will then use GPUs if at least one is found,
or CPUs otherwise. For all tests were a type of device is relevant,
command line options have been added to the CMakeLists files.