Fix bug in retrieving number of computing units
The clGetDeviceInfo function was called to retrieve the
CL_DEVICE_MAX_COMPUTE_UNITS value but the type of this parameter was
set to size_t instead of to cl_uint as defined in the standard. Not
complying with this aspect was (sometimes) leading to wrong numbers
which in turn might trigger e.g. errors in creating OpenCL buffers
because of trying to reserve huge memory chunks (just because of
meaningless numbers).