Make gaugemomentum load/dump methods as other ones
These methods for the gaugemomentum were using the import/export ones
of the hardware::code corresponding class, which in turn were doing the
conversion with the additional, needed Plain buffer. This
responsibility belongs to the buffer class and it is now done as it
should and as it is done for the other buffer objects.