Make explicit tests per device type for buffers
The hardware::buffers code contains Soa versus AoS branches and it is
meaningful to implement ctest test cases that are explicitly meant for
CPUs or GPUs. This has been done and the scenario were a given type of
device is not available has been handled making the test pass as it had
been done for the tests in hardware::code.