Make all inline OpenCL functions static inline
OpenCL C language is based on C99, so it uses the C99 semantics (which
are different than the C89 or C++ semantics) for the inline
keyword. For C99 static inline is the proper declaration for functions
that should be inlined in the current translation unit. More precisily,
in C99, inline means that a function's definition is provided only for
inlining, and that there is another definition (without inline)
somewhere else in the program. That means that a program with just a
function declared as inline is incomplete, because if such a function
is not inlined (for example, when compiling without optimization), then
there will be an unresolved reference to that other definition. This
turned out to be the case on AMD MI50 Instict GPUs using the ROCm
framework and it was then worth to fix this problem everyehere. It is
also worth to notice that in our way of building OpenCL code there is
always only one translation unit, since all code relevant for a given
kernel is prepared and put in a single file that is than build via the
clBuildProgram function.