Execute debug operation only if on single device
Some debug code in the method fermion_force_eo_device of
hardware::code::Molecular_Dynamics was using the polyakov_device method
of gaugefield which is meant to be used only on single device. However,
this was not enforced triggering some tests failures in physics in
environment with multiple device available.